Stairway Scene  GUARDROOM Special door
I had new kind of door with the elevator door mechanism

+++ TFBS/Assets/Scripts/Doors/ElevatorDoor.cs
@@ -2,9 +2,9 @@
 
 public class ElevatorDoor : MonoBehaviour
 {
-    public int XOffsetFromCenter;
-    public int YOffsetFromCenter;
-    public int ZOffsetFromCenter;
+    public float XOffsetFromCenter;
+    public float YOffsetFromCenter;
+    public float ZOffsetFromCenter;
     bool isPlayerNear = false;
     float animationProgress = 0.0f;
